refactor: 주문 세부 조회시 역할별 분기 로직을 전략 패턴으로 분리, 데이터 없을 때 널 처리
